improve example
